[chore][pkg/stanza] Cleanup syslog input operator files
