+++ pkg/stanza/operator/input/syslog/input_test.go
